backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
python-dotenv==1.0.1
google-genai>=1.10.0
google-adk==1.1.1
google-cloud-aiplatform==1.140.0
aiosqlite==0.20.0
pydantic==2.10.4
Pillow==11.1.0
python-multipart==0.0.20
